+++ vidl/__init__.py
@@ -27,7 +27,6 @@ def show_help():
     print(        '    '+cyan(package_name)+' [format] [options] <URL>')
     print(        '')
     print(green(  'Download Options:'))
-    print(cyan(   '    format             ')+'mp3, mp4, wav or m4a. Default mp3')
     print(cyan(   '    format             ')+'bestvideo, bestaudio, mp3, mp4, wav or m4a. Default mp3')
     print(cyan(   '    --no-md            ')+'Don\'t add metadata to downloaded files')
     print(cyan(   '    --no-smart-md      ')+'Don\'t extract artist and song name from title')